Moves MoveServer() to ITGAdministration
