+++ TGServerService/Config.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using TGServiceInterface;
+using System.ServiceModel;
 using System.Threading;
+using TGServiceInterface;
 
 namespace TGServerService
 {
@@ -288,119 +289,6 @@ namespace TGServerService
 			}
 		}
 
-		//public api
-		public string MoveServer(string new_location)
-		{
-			var Config = Properties.Settings.Default;
-			try
-			{
-				var di1 = new DirectoryInfo(Config.ServerDirectory);
-				var di2 = new DirectoryInfo(new_location);
-
-				var copy = di1.Root.FullName != di2.Root.FullName;
-
-				if (copy && File.Exists(PrivateKeyPath))
-					return String.Format("Unable to perform a cross drive server move with the {0}. Copy aborted!", PrivateKeyPath);
-
-				new_location = di2.FullName;
-
-				while (di2.Parent != null)
-					if (di2.Parent.FullName == di1.FullName)
-						return "Cannot move to child of current directory!";
-					else
-						di2 = di2.Parent;
-
-				if (!Monitor.TryEnter(RepoLock))
-					return "Repo locked!";
-				try
-				{
-					if (RepoBusy)
-						return "Repo busy!";
-					DisposeRepo();
-					if (!Monitor.TryEnter(ByondLock))
-						return "BYOND locked";
-					try
-					{
-						if (updateStat != TGByondStatus.Idle)
-							return "BYOND busy!";
-						if (!Monitor.TryEnter(CompilerLock))
-							return "Compiler locked!";
-
-						try
-						{
-							if (compilerCurrentStatus != TGCompilerStatus.Uninitialized && compilerCurrentStatus != TGCompilerStatus.Initialized)
-								return "Compiler busy!";
-							if (!Monitor.TryEnter(watchdogLock))
-								return "Watchdog locked!";
-							try
-							{
-								if (currentStatus != TGDreamDaemonStatus.Offline)
-									return "Watchdog running!";
-								lock (configLock)
-								{
-									CleanGameFolder();
-									Program.DeleteDirectory(GameDir);
-									string error = null;
-									if (copy)
-									{
-										Program.CopyDirectory(Config.ServerDirectory, new_location);
-										Directory.CreateDirectory(new_location);
-										Environment.CurrentDirectory = new_location;
-										try
-										{
-											Program.DeleteDirectory(Config.ServerDirectory);
-										}
-										catch (Exception e)
-										{
-											error = "The move was successful, but the path " + Config.ServerDirectory + " was unable to be deleted fully!";
-											TGServerService.WriteWarning(String.Format("Server move from {0} to {1} partial success: {2}", Config.ServerDirectory, new_location, e.ToString()), TGServerService.EventID.ServerMovePartial);
-										}
-									}
-									else
-									{
-										try
-										{
-											Environment.CurrentDirectory = di2.Root.FullName;
-											Directory.Move(Config.ServerDirectory, new_location);
-											Environment.CurrentDirectory = new_location;
-										}
-										catch
-										{
-											Environment.CurrentDirectory = Config.ServerDirectory;
-											throw;
-										}
-									}
-									TGServerService.WriteInfo(String.Format("Server moved from {0} to {1}", Config.ServerDirectory, new_location), TGServerService.EventID.ServerMoveComplete);
-									Config.ServerDirectory = new_location;
-									return null;
-								}
-							}
-							finally
-							{
-								Monitor.Exit(watchdogLock);
-							}
-						}
-						finally
-						{
-							Monitor.Exit(CompilerLock);
-						}
-					}
-					finally
-					{
-						Monitor.Exit(ByondLock);
-					}
-				}
-				finally
-				{
-					Monitor.Exit(RepoLock);
-				}
-			}
-			catch (Exception e)
-			{
-				TGServerService.WriteError(String.Format("Server move from {0} to {1} failed: {2}", Config.ServerDirectory, new_location, e.ToString()), TGServerService.EventID.ServerMoveFailed);
-				return e.ToString();
-			}
-		}
 		//public api
 		public ushort InteropPort(out string error)
 		{
@@ -816,7 +704,7 @@ namespace TGServerService
 				return e.ToString();
 			}
 		}
-
+		[OperationBehavior(Impersonation = ImpersonationOption.Required)]
 		public string ReadRaw(string configRelativePath, bool repo, out string error)
 		{
 			try
@@ -854,7 +742,7 @@ namespace TGServerService
 				return null;
 			}
 		}
-
+		[OperationBehavior(Impersonation = ImpersonationOption.Required)]
 		public string WriteRaw(string configRelativePath, string data)
 		{
 			try
@@ -990,6 +878,7 @@ namespace TGServerService
 			ranks.Remove(rank);
 			return WriteAdminRanks(ranks);
 		}
+
 		public string SetAdminRank(string rankName, IDictionary<string, bool> permissions)
 		{
 			var ranks = AdminRanks(out string error);
